src/third_party/blink/web_tests/fast/dom/HTMLImageElement/image-picture-source-src-expected.txt

CONSOLE WARNING: line 8: <source src> with a <picture> parent is invalid and therefore ignored. Please use <source srcset> instead.
